+++ src/objects/EquirectCamera.js
