+++ src/objects/PhysicalSpotLight.js
