+++ src/objects/ShapedAreaLight.js
